requirements.txt
amqp==5.2.0
asgiref==3.7.2
astroid==3.0.2
async-timeout==3.0.1
billiard==4.2.0
black==23.12.1
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dill==0.3.7
Django==5.0.1
django-countries==7.5.1
django-redis==5.4.0
djangorestframework==3.14.0
drf-yasg==1.21.7
future==0.18.3
greenlet==3.0.3
inflection==0.5.1
isort==5.13.2
kombu==5.3.4
mccabe==0.7.0
mypy==1.8.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
prompt-toolkit==3.0.43
psycopg==3.1.17
psycopg2-binary==2.9.9
pylint==3.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
serializers==0.2.4
six==1.16.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
tomlkit==0.12.3
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
